[Issue - #46] Widen transactions.amount to bigint
int4 milliunits cap a single transaction at 2,147,483.647 ARS. At the BCRA
reference rate that is roughly USD 1,400 — below a month's rent, a medical
bill, or a used-car payment — so the cap is a product defect for an
Argentine finance app, not a theoretical limit.

Landed ahead of the #46 handlers deliberately: implementing transactions
against a cap we already knew was wrong would have meant writing throwaway
int32 bounds checks and tests, then deleting them.

Shipped as one change because any partial version leaves two different caps
in one API:

- Migration 00005 widens the column, append-only. Its Down deliberately
  fails if a row no longer fits in int4 rather than truncating financial
  data during a rollback.
- The jsonb_to_recordset column list in BulkCreateTransactions moves to
  bigint. Left at integer, bulk-create and the entire CSV import path would
  have kept the old cap while single create no longer had it — and the
  import path is the one users hit with a year of history.
- Every OpenAPI milliunit field (TransactionInput, Transaction,
  TransactionListItem, SummaryCategory.value, SummaryDay.income/expenses,
  Summary.*Amount) becomes format: int64 bounded to ±(2^53-1). Go and TS
  types and the sqlc models regenerate from it.
- Drizzle moves to bigint('amount', { mode: 'number' }). This is required,
  not cosmetic: node-postgres returns int8 as a string, so without it the
  legacy stack would start receiving "12500" and break its arithmetic and
  formatting.
- The transaction form and CSV importer guard with isSafeMiliunitAmount.

The API bound is narrower than the column on purpose: ±(2^53-1) milliunits
is JavaScript's safe-integer limit, so every value the API returns is exact
in the browser. Values between that and bigint's range are reachable only by
direct SQL.

Note the contract's minimum/maximum are documentation only — oapi-codegen
emits no range validation — so the #46 handlers must check the bound
explicitly.

Registry 0006 is closed out and the schema test now asserts int8.
TestMigration00005_PreservesAmountsAndLiftsInt32Cap_LiveDatabase seeds both
int4 extremes plus an ordinary pair and zero, upgrades, and proves every
value survives sign-intact, then inserts 3,000,000,000 milliunits to prove
the cap is actually gone.

Verification: go build, go vet, full backend suite green against live
PostgreSQL; openapi:validate 28 operations; bun lint clean; bun test 17/17.

+++ backend/internal/db/queries/transactions.sql
@@ -92,7 +92,12 @@ INSERT INTO transactions (id, amount, payee, notes, date, account_id, category_i
 SELECT r.id, r.amount, r.payee, r.notes, r.date, r.account_id, r.category_id, r.currency
 FROM jsonb_to_recordset(sqlc.arg(payload)::jsonb) AS r(
     id text,
-    amount integer,
+    -- Must track transactions.amount's column type (bigint since migration
+    -- 00005). Left as integer, this recordset column would raise "value out
+    -- of range for type integer" for any amount past the old int4 cap, so
+    -- bulk-create and the CSV import path would keep a limit that single
+    -- create no longer has — two different caps in one API.
+    amount bigint,
     payee text,
     notes text,
     date timestamp,
